auth-backend: cleanup types and move them closer to home
