+++ plugins/auth-backend/src/providers/types.ts
@@ -19,21 +19,6 @@ import { Logger } from 'winston';
 import { TokenIssuer } from '../identity';
 import { Config } from '@backstage/config';
 
-export type OAuthProviderOptions = {
-  /**
-   * Client ID of the auth provider.
-   */
-  clientId: string;
-  /**
-   * Client Secret of the auth provider.
-   */
-  clientSecret: string;
-  /**
-   * Callback URL to be passed to the auth provider to redirect to after the user signs in.
-   */
-  callbackUrl: string;
-};
-
 export type AuthProviderConfig = {
   /**
    * The protocol://domain[:port] where the app is hosted. This is used to construct the
@@ -46,49 +31,6 @@ export type AuthProviderConfig = {
    */
   appUrl: string;
 };
-
-/**
- * Any OAuth provider needs to implement this interface which has provider specific
- * handlers for different methods to perform authentication, get access tokens,
- * refresh tokens and perform sign out.
- */
-export interface OAuthProviderHandlers {
-  /**
-   * This method initiates a sign in request with an auth provider.
-   * @param {express.Request} req
-   * @param options
-   */
-  start(
-    req: express.Request,
-    options: Record<string, string>,
-  ): Promise<RedirectInfo>;
-
-  /**
-   * Handles the redirect from the auth provider when the user has signed in.
-   * @param {express.Request} req
-   */
-  handler(
-    req: express.Request,
-  ): Promise<{
-    response: AuthResponse<OAuthProviderInfo>;
-    refreshToken?: string;
-  }>;
-
-  /**
-   * (Optional) Given a refresh token and scope fetches a new access token from the auth provider.
-   * @param {string} refreshToken
-   * @param {string} scope
-   */
-  refresh?(
-    refreshToken: string,
-    scope: string,
-  ): Promise<AuthResponse<OAuthProviderInfo>>;
-
-  /**
-   * (Optional) Sign out of the auth provider.
-   */
-  logout?(): Promise<void>;
-}
 
 /**
  * Any Auth provider needs to implement this interface which handles the routes in the
@@ -180,8 +122,6 @@ export type AuthResponse<ProviderInfo> = {
   backstageIdentity?: BackstageIdentity;
 };
 
-export type OAuthResponse = AuthResponse<OAuthProviderInfo>;
-
 export type BackstageIdentity = {
   /**
    * The backstage user ID.
@@ -192,67 +132,6 @@ export type BackstageIdentity = {
    * An ID token that can be used to authenticate the user within Backstage.
    */
   idToken?: string;
-};
-
-export type OAuthProviderInfo = {
-  /**
-   * An access token issued for the signed in user.
-   */
-  accessToken: string;
-  /**
-   * (Optional) Id token issued for the signed in user.
-   */
-  idToken?: string;
-  /**
-   * Expiry of the access token in seconds.
-   */
-  expiresInSeconds?: number;
-  /**
-   * Scopes granted for the access token.
-   */
-  scope: string;
-  /**
-   * A refresh token issued for the signed in user
-   */
-  refreshToken?: string;
-};
-
-export type OAuthPrivateInfo = {
-  /**
-   * A refresh token issued for the signed in user.
-   */
-  refreshToken: string;
-};
-
-/**
- * Payload sent as a post message after the auth request is complete.
- * If successful then has a valid payload with Auth information else contains an error.
- */
-export type WebMessageResponse =
-  | {
-      type: 'authorization_response';
-      response: AuthResponse<unknown>;
-    }
-  | {
-      type: 'authorization_response';
-      error: Error;
-    };
-
-export type PassportDoneCallback<Res, Private = never> = (
-  err?: Error,
-  response?: Res,
-  privateInfo?: Private,
-) => void;
-
-export type RedirectInfo = {
-  /**
-   * URL to redirect to
-   */
-  url: string;
-  /**
-   * Status code to use for the redirect
-   */
-  status?: number;
 };
 
 /**
@@ -276,39 +155,3 @@ export type ProfileInfo = {
    */
   picture?: string;
 };
-
-export type RefreshTokenResponse = {
-  /**
-   * An access token issued for the signed in user.
-   */
-  accessToken: string;
-  /**
-   * Optionally, the server can issue a new Refresh Token for the user
-   */
-  refreshToken?: string;
-  params: any;
-};
-
-export type ProviderStrategy = {
-  userProfile(accessToken: string, callback: Function): void;
-};
-
-export type SAMLProviderConfig = {
-  entryPoint: string;
-  issuer: string;
-};
-
-export type SAMLEnvironmentProviderConfig = {
-  [key: string]: SAMLProviderConfig;
-};
-
-export type OAuthState = {
-  /* A type for the serialized value in the `state` parameter of the OAuth authorization flow
-   */
-  nonce: string;
-  env: string;
-};
-
-export type EnvironmentIdentifierFn = (
-  req: express.Request,
-) => string | undefined;
